refactor: externald data movido
