+++ src/UI/orden/MobileOrderInfoItem.tsx
@@ -1,4 +1,4 @@
-import { ExtendedOrdenData } from '@utils/Examples/BasicOrderTable'
+import { ExtendedOrdenData } from "@utils/Examples/ExtendedOrdenData"
 import Image from 'next/image'
 import Link from 'next/link'
 import React from 'react'
